pyproject.toml
[tool.poetry]
name = "airflow-task-pytest"
version = "0.1.0"
description = ""
authors = ["choiwonil <eywanna@g.skku.edu>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
apache-airflow = "2.1.2"
moto = {extras = ["s3"], version = "2.2.2"}
awswrangler = "2.10.0"
pytest = "6.2.4"
python-dotenv = "0.19.0"
pandas = "1.3.1"
pytest-env = "0.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
